requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.9
dj-database-url==2.3.0
